requirements.txt
matplotlib~=3.3.4
numpy~=1.23.4
scikit-image~=0.19.3
pydicom~=2.3.0
gdcm
pylibjpeg-libjpeg
pyyaml~=5.3.1
pytorch-ignite~=0.4.10
opencv-python~=4.6.0.66
SimpleITK~=2.2.0
tqdm~=4.61.1
Pillow~=9.3.0
torchvision~=0.10.0
torch~=1.13.0
imageio~=2.9.0
setuptools~=60.2.0
ipython~=8.6.0
scipy~=1.9.3
scikit-learn~=1.1.3
pyhocon~=0.3.59
dotmap~=1.3.30
ipdb~=0.13.9
lpips~=0.1.4